requirements.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain-community
langchain-core
langgraph
ddgs
streamlit
httpx
pydantic